pyproject.toml
[tool.poetry]
name = "platzi"
version = "0.7.3"
description = "Download your courses from platzi"
authors = ["ivansaul <ivansaul@no-reply.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ivansaul/platzi-downloader"
repository = "https://github.com/ivansaul/platzi-downloader"
keywords = ["Platzi", "downloader", "courses"]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

packages = [
    { include = "scraper" },
    { include = "backend" }
]

[tool.poetry.scripts]
platzi = "scraper.api:AsyncPlatzi"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.19.2"
pydantic = "^2.9.2"
pyhumps = "^3.8.0"
tqdm = "^4.67.0"
pytest-playwright = "^0.5.2"
unidecode = "^1.3.8"
platformdirs = "^4.3.6"
rnet = "^2.4.2"
aiofiles = "^25.1.0"
fastapi = "^0.135.1"
uvicorn = "^0.42.0"
websockets = "^16.0"
jinja2 = "^3.1.6"
httpx = "^0.28.1"
google-api-python-client = "*"
google-auth-oauthlib = "*"
google-auth-httplib2 = "*"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.14.1"
ipykernel = "^6.29.5"
python-semantic-release = "^9.14.0"
mypy = "^1.13.0"
types-aiofiles = "^25.1.0.20251011"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
preview = true
select = ['E', 'F', 'W', 'I']
ignore = ["E501"]

[tool.semantic_release]
version_source = "tag"
version_toml = ["pyproject.toml:tool.poetry.version"]
branch = "master"
build_command = "pip install poetry && poetry build"
major_on_zero = true
allow_zero_version = true
upload_to_pypi = true
upload_to_release = true
commit_version_number = true
